deleted snapshots.ts and its reference in background.js. Moved WebMetricsContainer to containers folder
