--- src/app/components/snapshots.ts
@@ -1,150 +1,0 @@
-const snapshots: {
-  snapshot: number;
-  component1: number;
-  component2: number;
-  component3: number;
-  [key: string]: number;
-}[] = [
-  {
-    snapshot: 10,
-    component1: 32,
-    component2: 27,
-    component3: 12,
-    'all others': 14,
-  },
-  {
-    snapshot: 2,
-    component1: 73,
-    component2: 45,
-    component3: 23,
-    'all others': 80,
-  },
-  {
-    snapshot: 3,
-    component1: 84,
-    component2: 46,
-    component3: 15,
-    'all others': 46,
-  },
-  {
-    snapshot: 4,
-    component1: 78,
-    component2: 45,
-    component3: 36,
-    'all others': 10,
-  },
-  {
-    snapshot: 5,
-    component1: 59,
-    component2: 52,
-    component3: 49,
-    'all others': 28,
-  },
-  {
-    snapshot: 6,
-    component1: 36,
-    component2: 96,
-    component3: 1,
-    'all others': 46,
-  },
-  {
-    snapshot: 7,
-    component1: 6,
-    component2: 6,
-    component3: 42,
-    'all others': 99,
-  },
-  {
-    snapshot: 8,
-    component1: 76,
-    component2: 29,
-    component3: 44,
-    'all others': 47,
-  },
-  {
-    snapshot: 9,
-    component1: 43,
-    component2: 62,
-    component3: 5,
-    'all others': 41,
-  },
-  {
-    snapshot: 10,
-    component1: 10,
-    component2: 84,
-    component3: 54,
-    'all others': 91,
-  },
-  {
-    snapshot: 11,
-    component1: 5,
-    component2: 78,
-    component3: 38,
-    'all others': 8,
-  },
-  {
-    snapshot: 12,
-    component1: 67,
-    component2: 73,
-    component3: 31,
-    'all others': 50,
-  },
-  {
-    snapshot: 13,
-    component1: 43,
-    component2: 42,
-    component3: 7,
-    'all others': 28,
-  },
-  {
-    snapshot: 14,
-    component1: 48,
-    component2: 26,
-    component3: 1,
-    'all others': 30,
-  },
-  {
-    snapshot: 15,
-    component1: 60,
-    component2: 28,
-    component3: 64,
-    'all others': 69,
-  },
-  {
-    snapshot: 16,
-    component1: 70,
-    component2: 15,
-    component3: 32,
-    'all others': 91,
-  },
-  {
-    snapshot: 17,
-    component1: 34,
-    component2: 86,
-    component3: 46,
-    'all others': 61,
-  },
-  {
-    snapshot: 18,
-    component1: 7,
-    component2: 50,
-    component3: 26,
-    'all others': 32,
-  },
-  {
-    snapshot: 19,
-    component1: 70,
-    component2: 40,
-    component3: 74,
-    'all others': 6,
-  },
-  {
-    snapshot: 20,
-    component1: 33,
-    component2: 78,
-    component3: 59,
-    'all others': 91,
-  },
-];
-
-export default snapshots;
